+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/note/types/GitDiffView.kt
@@ -60,7 +60,7 @@ import com.vitorpamplona.amethyst.R
 import com.vitorpamplona.amethyst.commons.icons.symbols.Icon
 import com.vitorpamplona.amethyst.commons.icons.symbols.MaterialSymbol
 import com.vitorpamplona.amethyst.commons.icons.symbols.MaterialSymbols
-import com.vitorpamplona.amethyst.ui.screen.loggedIn.gitRepo.code.CodeHighlighter
+import com.vitorpamplona.amethyst.commons.nip34Git.ui.CodeHighlighter
 import com.vitorpamplona.amethyst.ui.stringRes
 import com.vitorpamplona.quartz.nip34Git.patch.CharSpan
 import com.vitorpamplona.quartz.nip34Git.patch.GitDiffFile
