+++ commons/src/jvmAndroid/kotlin/com/vitorpamplona/amethyst/commons/nip34Git/GitRepositoryBrowserViewModel.kt
@@ -18,7 +18,7 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.ui.screen.loggedIn.gitRepo.code
+package com.vitorpamplona.amethyst.commons.nip34Git
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
